refactor: improve DTO naming and add delete validation

+++ src/main/java/p/projects/springbookstore/service/impl/BookServiceImpl.java
@@ -60,6 +60,9 @@ public class BookServiceImpl implements BookService {
     @Override
     @Transactional
     public void deleteBookById(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find book with id: " + id);
+        }
         bookRepository.deleteById(id);
     }
